# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Nexis License System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getAdminProfile
      tags: [auth]
      summary: Get current admin profile
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─────────────────────────────────────────────────────────────────
  # Public License Endpoints (used by Chrome Extension)
  # ─────────────────────────────────────────────────────────────────
...
  /admin/plans:
    get:
      operationId: listPlans
      tags: [admin-plans]
      summary: List all plans
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [admin-plans]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    get:
      operationId: getPlan
      tags: [admin-plans]
      summary: Get a plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    put:
      operationId: updatePlan
      tags: [admin-plans]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [admin-plans]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Users
  # ─────────────────────────────────────────────────────────────────
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin-users]
      summary: List users (license holders)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin-users]
      summary: Get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
    delete:
      operationId: deleteUser
      tags: [admin-users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Devices
  # ─────────────────────────────────────────────────────────────────
...
  /admin/notifications:
    get:
      operationId: listNotifications
      tags: [admin-notifications]
      summary: List notifications
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
    post:
      operationId: createNotification
      tags: [admin-notifications]
      summary: Create and broadcast a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Analytics
  # ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth ──────────────────────────────────────────────────────
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminProfile"
      required: [token, admin]
    AdminProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, username, email, createdAt]

    # ── License Public ─────────────────────────────────────────────
...
    License:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        status:
          type: string
          enum: [active, suspended, revoked, expired, trial]
        planId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        maxDevices:
          type: integer
        activatedDevices:
          type: integer
        expiresAt:
          type: ["string", "null"]
        activatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        notes:
          type: ["string", "null"]
      required: [id, key, status, maxDevices, activatedDevices, createdAt]
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        durationDays:
          type: ["integer", "null"]
        maxDevices:
          type: integer
        isActive:
          type: boolean
        licenseCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, price, currency, maxDevices, isActive, licenseCount, createdAt]
    PlanInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        durationDays:
          type: ["integer", "null"]
        maxDevices:
          type: integer
        isActive:
          type: boolean
      required: [name, price, maxDevices]
    PlanUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        durationDays:
          type: ["integer", "null"]
        maxDevices:
          type: ["integer", "null"]
        isActive:
          type: ["boolean", "null"]

    # ── Users ──────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        licenseCount:
          type: integer
        activeDevices:
          type: integer
        createdAt:
          type: string
        lastSeen:
          type: ["string", "null"]
      required: [id, licenseCount, activeDevices, createdAt]
    UserDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        licenseCount:
          type: integer
        activeDevices:
          type: integer
        createdAt:
          type: string
        lastSeen:
          type: ["string", "null"]
        licenses:
          type: array
          items:
            $ref: "#/components/schemas/License"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
      required: [id, licenseCount, activeDevices, createdAt, licenses, devices]
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ── Devices ────────────────────────────────────────────────────
    Device:
      type: object
      properties:
        id:
          type: integer
        licenseId:
          type: integer
        deviceId:
          type: string
        deviceInfo:
          type: ["object", "null"]
          additionalProperties: true
        firstSeen:
          type: string
        lastSeen:
          type: string
        isActive:
          type: boolean
      required: [id, licenseId, deviceId, firstSeen, lastSeen, isActive]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error, success]
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, title, message, type, isRead, createdAt]
    NotificationInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error, success]
      required: [title, message, type]
    NotificationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ── Analytics ─────────────────────────────────────────────────
...