# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexis License System API
 * OpenAPI spec version: 1.0.0
 */

export interface Plan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  currency: string;
  /** @nullable */
  durationDays?: number | null;
  maxDevices: number;
  isActive: boolean;
  licenseCount: number;
  createdAt: string;
}