# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexis License System API
 * OpenAPI spec version: 1.0.0
 */

export interface PlanUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  durationDays?: number | null;
  /** @nullable */
  maxDevices?: number | null;
  /** @nullable */
  isActive?: boolean | null;
}